Changing score to ulong in the unlikely event of a score cap

+++ RaylibDemo/player/PlayerData.cs
@@ -6,7 +6,7 @@ namespace PointDefence.Player
     {
         public static int health = 100;
         public static int ammo = 100;
-        public static int score = 0;
+        public static ulong score = 0;
 
         public static void ResetStats()
         {
